requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
feedparser==6.0.12
fonttools==4.60.2
idna==3.11
importlib_resources==6.5.2
joblib==1.5.3
kiwisolver==1.4.7
lxml==6.0.2
matplotlib==3.9.4
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pyarrow==21.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zipp==3.23.0
